src/requirements.txt
Flask==2.3.1
flask_api==3.1
Werkzeug==2.3.0
openai==1.65.4
python-dotenv==1.0.1
